+++ src/helpers.js
@@ -1,6 +1,11 @@
 const $path = require('path');
 const $util = require('util');
 
+/**
+ * Normalizes the path in a way we can handle windows and mac pathes nearly the same way
+ * 
+ * @param {string} path - A path that comes from VS code
+ */
 exports.normalizePath = (path) => {
 	// the uri.path in vscode is sometimes with forward slashes...
 	path = $path.normalize(path);
@@ -9,6 +14,34 @@ exports.normalizePath = (path) => {
 		return hits[1].toUpperCase();
 	});
 	return path;
+}
+
+/**
+ * Extracts a path string from a VS Code input object
+ * 
+ * @param {object} input - An input object (or document) that comes from VS code
+ */
+ exports.getPath = (input) => {
+	if  (typeof input.uri !== "undefined") {
+        return exports.normalizePath(input.uri.path);
+    // diff items
+    } else if (typeof input.original !== "undefined") {
+        return exports.normalizePath(input.original.path);
+    }
+}
+
+/**
+ * Generates an Id for the flat repository
+ * 
+ * @param {object} input - An input object (or document) that comes from VS code
+ */
+ exports.getId = (input) => {
+	if  (typeof input.uri !== "undefined") {
+        return exports.normalizePath(input.uri.path);
+    // diff items
+    } else if (typeof input.original !== "undefined") {
+        return `${exports.normalizePath(input.original.path)}|${exports.normalizePath(input.modified.path)}`;
+    }
 }
 
 exports.log = () => {