+++ CHANGELOG.md
@@ -7,3 +7,22 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ## [Unreleased]
 
 - Initial release
+
+## Release Notes
+
+### 1.0.0
+
+ * Initial release
+
+### 1.0.2
+
+ * Just some non code relevant corrections for the release
+
+### 1.0.3
+
+ * Fix: Settings icon is now only displayed on the Better Open Editors panel
+ * Fix: Gitlens and standard file diffs are now displayed
+ * Fix: New files support improved
+ * Fix: File/folder sorting improved
+
+